@@ -15,7 +15,7 @@
         @routes
         <script src="{{ mix('js/app.js') }}" defer></script>
     </head>
-    <body>
+    <body style="max-width: 720px; margin-left:auto; margin-right: auto;">
         @inertia
 
         @env ('local')
